Split coordinate engine into a package
